refactor amp embed cleaner to be more readable and to be able to cope with element-embed types as well
